pet_express_scraper/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
requests>=2.31.0
beautifulsoup4>=4.12.0
html5lib>=1.1
httpx>=0.24.0
python-dotenv>=1.0.0
pydantic-ai>=0.0.35
tavily-python>=0.5.1
openai>=1.13.0
aiohttp>=3.9.0
langchain>=0.1.0
langchain-openai>=0.0.5
# Using direct API calls instead of JavaScript rendering
